backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.116.0
uvicorn==0.35.0
python-multipart==0.0.20
starlette==0.46.2

# Database dependencies
sqlalchemy==2.0.41
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Authentication and security
pydantic==2.11.7
email-validator==2.2.0
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
cryptography==45.0.5

# Environment and configuration
python-dotenv==1.1.1

# Supporting libraries
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
pyasn1==0.6.1
pycparser==2.22
pydantic_core==2.33.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1

# Google Cloud Storage dependencies
google-cloud-storage==2.20.0
google-resumable-media==2.7.2
google-auth==2.35.0
google-cloud-core==2.4.1

# Video processing dependencies
opencv-python==4.10.0.84
Pillow==11.0.0
mediapipe==0.10.21

# Async file operations
aiofiles==24.1.0

# HTTP client for OAuth
httpx==0.27.0

# Google Gemini AI dependencies (new v2 API)
google-genai

# OpenAI dependencies for TTS
openai==1.35.0